:truck: Renamed social_media to social_network
