+++ social_network/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class SocialMediaConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "social_media"
+    name = "social_network"
