+++ social_network_api_service/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for social_media_api_service project.
+ASGI config for social_network_api_service project.
 """
 
 import os
@@ -8,7 +8,7 @@ from django.core.asgi import get_asgi_application
 
 os.environ.setdefault(
     "DJANGO_SETTINGS_MODULE",
-    "social_media_api_service.settings"
+    "social_network_api_service.settings"
 )
 
 application = get_asgi_application()
