+++ social_network_api_service/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for social_media_api_service project.
+WSGI config for social_network_api_service project.
 """
 
 import os
@@ -8,7 +8,7 @@ from django.core.wsgi import get_wsgi_application
 
 os.environ.setdefault(
     "DJANGO_SETTINGS_MODULE",
-    "social_media_api_service.settings"
+    "social_network_api_service.settings"
 )
 
 application = get_wsgi_application()
